chore: move `RestaurantDetailsScreen` file out of the folder
